move mock client to seperate file
