+++ pkg/client_wrapper/client_wrapper.go
@@ -10,19 +10,3 @@ type ClientFunctions interface {
 	Subscribed(ctx context.Context, opts *reddit.ListSubredditOptions) ([]*reddit.Subreddit, *reddit.Response, error)
 	Upvoted(ctx context.Context, opts *reddit.ListUserOverviewOptions) ([]*reddit.Post, *reddit.Response, error)
 }
-
-type RedditClient struct {
-	Client *reddit.Client
-}
-
-var ctx = context.Background()
-
-func (r *RedditClient) Subscribed(ctx context.Context, opts *reddit.ListSubredditOptions) ([]*reddit.Subreddit, *reddit.Response, error) {
-	s, res, err := r.Client.Subreddit.Subscribed(ctx, opts)
-	return s, res, err
-}
-
-func (r *RedditClient) Upvoted(ctx context.Context, opts *reddit.ListUserOverviewOptions) ([]*reddit.Post, *reddit.Response, error) {
-	s, res, err := r.Client.User.Upvoted(ctx, opts)
-	return s, res, err
-}
