+++ pkg/subreddits/subreddits_test.go
@@ -1,32 +1,12 @@
 package subreddits
 
 import (
-	"context"
 	"reflect"
 	"testing"
 
 	"github.com/clemak27/greddit/pkg/client_wrapper"
 	"github.com/vartanbeno/go-reddit/v2/reddit"
 )
-
-type mockClient struct {
-	Client *reddit.Client
-}
-
-func (r *mockClient) Subscribed(ctx context.Context, opts *reddit.ListSubredditOptions) ([]*reddit.Subreddit, *reddit.Response, error) {
-	s := []*reddit.Subreddit{
-		{
-			ID:     "1",
-			FullID: "111",
-			Name:   "My awesome subreddit",
-		},
-	}
-	return s, nil, nil
-}
-
-func (r *mockClient) Upvoted(ctx context.Context, opts *reddit.ListUserOverviewOptions) ([]*reddit.Post, *reddit.Response, error) {
-	return nil, nil, nil
-}
 
 func TestGetSubscriptions(t *testing.T) {
 	type args struct {
@@ -41,7 +21,7 @@ func TestGetSubscriptions(t *testing.T) {
 		{
 			name: "GetSubscriptions is successfull",
 			args: args{
-				rc: &mockClient{},
+				rc: &client_wrapper.MockClient{},
 			},
 			wantL: []*reddit.Subreddit{
 				{
