+++ pkg/upvoted/upvoted_test.go
@@ -1,33 +1,12 @@
 package upvoted
 
 import (
-	"context"
 	"reflect"
 	"testing"
 
 	"github.com/clemak27/greddit/pkg/client_wrapper"
 	"github.com/vartanbeno/go-reddit/v2/reddit"
 )
-
-type mockClient struct {
-	Client *reddit.Client
-}
-
-func (r *mockClient) Subscribed(ctx context.Context, opts *reddit.ListSubredditOptions) ([]*reddit.Subreddit, *reddit.Response, error) {
-	return nil, nil, nil
-}
-
-func (r *mockClient) Upvoted(ctx context.Context, opts *reddit.ListUserOverviewOptions) ([]*reddit.Post, *reddit.Response, error) {
-	u := []*reddit.Post{
-		{
-			ID:     "1",
-			FullID: "111",
-			URL:    "localhost:8080",
-			Title:  "testPost",
-		},
-	}
-	return u, nil, nil
-}
 
 func TestGetUpvoted(t *testing.T) {
 	type args struct {
@@ -42,7 +21,7 @@ func TestGetUpvoted(t *testing.T) {
 		{
 			name: "GetUpvoted is successfull",
 			args: args{
-				rc: &mockClient{},
+				rc: &client_wrapper.MockClient{},
 			},
 			wantL: []*reddit.Post{
 				{
